delete task by id of the authorized user

+++ src/tasks/tasks.service.ts
@@ -45,15 +45,15 @@ export class TasksService {
      * @param id id in the database
      * @param status TaskStatus to update
      */
-    // async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    //     const task = await this.getTaskById(id);
-    //     task.status = status;
-    //     await task.save();
-    //     return task;
-    // }
+    async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user);
+        task.status = status;
+        await task.save();
+        return task;
+    }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(id: number, user: User): Promise<void> {
+        const result = await this.taskRepository.delete({ id, userId: user.id });
 
         if (result.affected === 0) {
             throw new NotFoundException(`Task with ID: "${id}" not found.`)
